Move shared conflicts find code to `passes/common`
Also refactor the `common` module and split by functionality area.
